+++ packages/components/frameworks/react-ssr/src/App.js
@@ -254,8 +254,13 @@ function App() {
 
   return (
     <div className="App">
-      <Accordion labelPosition="center" openLabel="open" closeLabel="close" content="TextContent"></Accordion>
       <h1>React preview</h1>
+      <Pagination
+        numberOfElements={100}
+        lastNumberLimit={99}
+        valueOnChange={(event) => console.log(event)}
+        dropdownMenuPos="top"
+      ></Pagination>
       <h2>Elvia ICONS</h2>
       <div className="testingIconSizes">
         <Icon name="folderCreate" customSize="37px"></Icon>